Code cleanups and README updates
